+++ modules/src/mock/client_def.rs
@@ -1,5 +1,9 @@
+use crate::ics02_client::client_consensus::AnyConsensusState;
+use crate::ics02_client::client_def::ClientDef;
+use crate::ics02_client::client_state::AnyClientState;
 use crate::ics03_connection::connection::ConnectionEnd;
 use crate::ics04_channel::channel::ChannelEnd;
+use crate::ics04_channel::packet::Sequence;
 use crate::ics23_commitment::commitment::{CommitmentPrefix, CommitmentProofBytes, CommitmentRoot};
 use crate::ics23_commitment::merkle::apply_prefix;
 use crate::ics24_host::identifier::{ChannelId, ClientId, ConnectionId, PortId};
@@ -7,10 +11,6 @@ use crate::ics24_host::Path;
 use crate::mock::client_state::{MockClientState, MockConsensusState};
 use crate::mock::header::MockHeader;
 use crate::Height;
-use crate::{
-    ics02_client::client_def::{AnyClientState, AnyConsensusState, ClientDef},
-    ics04_channel::packet::Sequence,
-};
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct MockClient;